refactor: move AssetAtlasException into Exceptions namespace
